Revert "Moved ListViewSubItem"
This reverts commit db3678c86f4d5ab3c8b9d34b4cbe2e5ee953cb1f.
